apa ini banyak banget pokoknya

+++ resources/views/panel/customers/index.blade.php
@@ -29,29 +29,28 @@
         <div class="card">
             <div class="card-body">
                 <div class="table-responsive overflow-auto">
-                    @if (Auth::user()->role == 'admin' || Auth::user()->role == 4)
-                        <div class="d-flex justify-content-between gap-2 align-items-center">
-                            <div class="d-flex gap-2">
-                                <a href="{{ route('customers.create') }}" class="btn btn-sm btn-primary">
-                                    Tambah Data
-                                </a>
-                                <div class="dropdown">
-                                    <button class="btn btn-sm btn-primary text-white" type="button"
-                                        data-bs-toggle="dropdown">
-                                        Export Data
-                                    </button>
-                                    <ul class="dropdown-menu">
-                                        <li><a class="dropdown-item" onclick="CopyToClipboard()">Copy</a></li>
-                                        <li><a class="dropdown-item" onclick="ExportToCSV()">CSV</a></li>
-                                        <li><a class="dropdown-item" onclick="ExportToPDF()">PDF</a></li>
-                                        <li><a class="dropdown-item" onclick="ExportToXLSX()">XLSX</a></li>
-                                    </ul>
-                                </div>
+                    <div class="d-flex justify-content-between gap-2 align-items-center">
+                        <div class="d-flex gap-2">
+                            <a href="{{ route('customers.create') }}"
+                                class="btn btn-sm btn-primary {{ Auth::user()->role == 'admin' || Auth::user()->role == 4 ? '' : 'd-none' }}">
+                                Tambah Data
+                            </a>
+                            <div
+                                class="dropdown {{ Auth::user()->role == 'admin' || Auth::user()->role == 4 || Auth::user()->role == 5 ? '' : 'd-none' }}">
+                                <button class="btn btn-sm btn-primary text-white" type="button" data-bs-toggle="dropdown">
+                                    Export Data
+                                </button>
+                                <ul class="dropdown-menu">
+                                    <li><a class="dropdown-item" onclick="CopyToClipboard()">Copy</a></li>
+                                    <li><a class="dropdown-item" onclick="ExportToCSV()">CSV</a></li>
+                                    <li><a class="dropdown-item" onclick="ExportToPDF()">PDF</a></li>
+                                    <li><a class="dropdown-item" onclick="ExportToXLSX()">XLSX</a></li>
+                                </ul>
                             </div>
-                            <a href="{{ route('customers.trash') }}" class="btn btn-sm btn-primary">Lihat Data yang
-                                Dihapus</a>
                         </div>
-                    @endif
+                        <a href="{{ route('customers.trash') }}" class="btn btn-sm btn-primary">Lihat Data yang
+                            Dihapus</a>
+                    </div>
                     <table class="table nowrap table-striped table-hover align-middle" id="datatable">
                         <thead>
                             <tr>
@@ -59,9 +58,9 @@
                                 <th>No Daftar</th>
                                 <th>Nama Lengkap</th>
                                 <th>No Telepon</th>
-                                {{-- <th {{ Auth::user()->role == 'admin' ? '' : 'class="hidden"' }}>Status
-                                </th> --}}
-                                <th class="bg-white text-dark">Aksi</th>
+                                <th
+                                    class="bg-white text-dark {{ Auth::user()->role == 'admin' || Auth::user()->role == 4 ? '' : 'd-none' }}">
+                                    Aksi</th>
                             </tr>
                         </thead>
                         <tbody class="text-capitalize" id="tbody">
@@ -75,25 +74,27 @@
                                         {{ $item->status == 1 ? 'Aktif' : 'Tidak Aktif' }}
                                     </td> --}}
                                     <td class="bg-white text-dark">
-                                        <div class="d-flex gap-1">
-                                            <a href="{{ route('customers.show', encrypt($item->id)) }}"
-                                                class="btn btn-primary btn-sm" title="Lihat Detail">
-                                                <i class="bi bi-info-circle-fill"></i>
-                                            </a>
-                                            <a href="{{ route('customers.edit', encrypt($item->id)) }}"
-                                                class="btn btn-warning btn-sm" title="Edit">
-                                                <i class="bi bi-pencil-square"></i>
-                                            </a>
-                                            <form action="{{ route('customers.destroy', encrypt($item->id)) }}"
-                                                method="POST">
-                                                @csrf
-                                                @method('DELETE')
-                                                <button type="button" onclick="deleteData(this)"
-                                                    class="btn btn-outline-danger btn-sm" title="Hapus">
-                                                    <i class="bi bi-trash3-fill"></i>
-                                                </button>
-                                            </form>
-                                        </div>
+                                        @if (Auth::user()->role == 'admin' || Auth::user()->role == 4)
+                                            <div class="d-flex gap-1">
+                                                <a href="{{ route('customers.show', encrypt($item->id)) }}"
+                                                    class="btn btn-primary btn-sm" title="Lihat Detail">
+                                                    <i class="bi bi-info-circle-fill"></i>
+                                                </a>
+                                                <a href="{{ route('customers.edit', encrypt($item->id)) }}"
+                                                    class="btn btn-warning btn-sm" title="Edit">
+                                                    <i class="bi bi-pencil-square"></i>
+                                                </a>
+                                                <form action="{{ route('customers.destroy', encrypt($item->id)) }}"
+                                                    method="POST">
+                                                    @csrf
+                                                    @method('DELETE')
+                                                    <button type="button" onclick="deleteData(this)"
+                                                        class="btn btn-outline-danger btn-sm" title="Hapus">
+                                                        <i class="bi bi-trash3-fill"></i>
+                                                    </button>
+                                                </form>
+                                            </div>
+                                        @endif
                                     </td>
                                 </tr>
                             @endforeach
